Adding and fixing more warnings

+++ tests/AppTest.cpp
@@ -633,24 +633,24 @@ TEST_F(TApp, RequiredFlags) {
 
 TEST_F(TApp, CallbackFlags) {
 
-    int value = 0;
+    size_t value = 0;
 
     auto func = [&value](size_t x) { value = x; };
 
     app.add_flag_function("-v", func);
 
     run();
-    EXPECT_EQ(value, 0);
+    EXPECT_EQ(value, (size_t)0);
 
     app.reset();
     args = {"-v"};
     run();
-    EXPECT_EQ(value, 1);
+    EXPECT_EQ(value, (size_t)1);
 
     app.reset();
     args = {"-vv"};
     run();
-    EXPECT_EQ(value, 2);
+    EXPECT_EQ(value, (size_t)2);
 
     EXPECT_THROW(app.add_flag_function("hi", func), CLI::IncorrectConstruction);
 }
@@ -658,24 +658,24 @@ TEST_F(TApp, CallbackFlags) {
 #if __cplusplus >= 201402L
 TEST_F(TApp, CallbackFlagsAuto) {
 
-    int value = 0;
+    size_t value = 0;
 
     auto func = [&value](size_t x) { value = x; };
 
     app.add_flag("-v", func);
 
     run();
-    EXPECT_EQ(value, 0);
+    EXPECT_EQ(value, (size_t)0);
 
     app.reset();
     args = {"-v"};
     run();
-    EXPECT_EQ(value, 1);
+    EXPECT_EQ(value, (size_t)1);
 
     app.reset();
     args = {"-vv"};
     run();
-    EXPECT_EQ(value, 2);
+    EXPECT_EQ(value, (size_t)2);
 
     EXPECT_THROW(app.add_flag("hi", func), CLI::IncorrectConstruction);
 }
